+++ src/shared/components/case-viewer/services/case.resolver.spec.ts
@@ -1,8 +1,8 @@
 import { CaseResolver } from './case.resolver';
 import { Observable } from 'rxjs';
-import createSpyObj = jasmine.createSpyObj;
 import { CaseView } from '../../../domain';
-import { DraftService, AlertService } from '../../../services';
+import { AlertService, DraftService } from '../../../services';
+import createSpyObj = jasmine.createSpyObj;
 
 describe('CaseResolver', () => {
   describe('resolve()', () => {
@@ -10,13 +10,15 @@ describe('CaseResolver', () => {
     const PARAM_CASE_ID = CaseResolver.PARAM_CASE_ID;
 
     const CASE_ID = '42';
-    const CASE: CaseView = createSpyObj<any>('case', ['toString']);
-    const CASE_CACHED: CaseView = createSpyObj<any>('caseCached', ['toString']);
+    const CASE: CaseView = new CaseView();
+    CASE.case_id = 'CASE_ID_1';
+
+    const CASE_CACHED: CaseView = new CaseView();
+    CASE_CACHED.case_id = 'CACHED_CASE_ID_1';
     const CASE_OBS: Observable<CaseView> = Observable.of(CASE);
 
     let caseResolver: CaseResolver;
     let draftService: DraftService;
-
     let casesService: any;
     let caseService: any;
     let alertService: AlertService;
@@ -28,7 +30,6 @@ describe('CaseResolver', () => {
       caseService = createSpyObj('caseService', ['announceCase']);
       casesService = createSpyObj('casesService', ['getCaseViewV2']);
       draftService = createSpyObj('draftService', ['getDraft']);
-
       router = createSpyObj('router', ['navigate']);
       alertService = createSpyObj('alertService', ['success']);
       caseResolver = new CaseResolver(caseService, casesService, draftService, router, alertService);
@@ -49,13 +50,13 @@ describe('CaseResolver', () => {
       caseResolver
         .resolve(route)
         .then(caseData => {
-          expect(caseData).toBe(CASE);
+          expect(caseData).toEqual(CASE);
         });
 
       expect(casesService.getCaseViewV2).toHaveBeenCalledWith(CASE_ID);
       expect(route.paramMap.get).toHaveBeenCalledWith(PARAM_CASE_ID);
       // allows to access private cachedCaseView field
-      expect(caseResolver['cachedCaseView']).toBe(CASE);
+      expect(caseResolver['cachedCaseView']).toEqual(CASE);
     });
 
     it('should return cached case view when the route is a case view tab and cached view exists', () => {
@@ -94,12 +95,12 @@ describe('CaseResolver', () => {
       caseResolver
         .resolve(route)
         .then(caseData => {
-          expect(caseData).toBe(CASE);
+          expect(caseData).toEqual(CASE);
         });
 
           expect(casesService.getCaseViewV2).toHaveBeenCalledWith(CASE_ID);
       // allows to access private cachedCaseView field
-          expect(caseResolver['cachedCaseView']).toBe(CASE);
+      expect(caseResolver['cachedCaseView']).toEqual(CASE);
     });
 
     it('should return cached case view when the route is not the one for case view and cached view exists', () => {
@@ -119,7 +120,7 @@ describe('CaseResolver', () => {
           expect(caseData).toBe(CASE_CACHED);
         });
       expect(casesService.getCaseViewV2).not.toHaveBeenCalled();
-      expect(caseResolver['cachedCaseView']).toBe(CASE_CACHED);
+      expect(caseResolver['cachedCaseView']).toEqual(CASE_CACHED);
     });
 
     it('should retrieve case view when the route is not the one for case view and cached is empty', () => {
@@ -136,12 +137,12 @@ describe('CaseResolver', () => {
       caseResolver
         .resolve(route)
         .then(caseData => {
-          expect(caseData).toBe(CASE);
+          expect(caseData).toEqual(CASE);
         });
 
       expect(casesService.getCaseViewV2).toHaveBeenCalledWith(CASE_ID);
       // allows to access private cachedCaseView field
-      expect(caseResolver['cachedCaseView']).toBe(CASE);
+      expect(caseResolver['cachedCaseView']).toEqual(CASE);
     });
 
     it('should redirect to error page when case cannot be retrieved', () => {
@@ -178,8 +179,11 @@ describe('CaseResolver', () => {
     const PARAM_CASE_ID = CaseResolver.PARAM_CASE_ID;
 
     const DRAFT_ID = 'DRAFT42';
-    const DRAFT: CaseView = createSpyObj<any>('draft', ['toString']);
-    const DRAFT_CACHED: CaseView = createSpyObj<any>('draftCached', ['toString']);
+    const DRAFT: CaseView = new CaseView();
+    DRAFT.case_id = 'DRAFT_CASE_ID_1';
+
+    const DRAFT_CACHED: CaseView = new CaseView();
+    DRAFT_CACHED.case_id = 'DRAFT_CASE_CACHED_ID_1';
     const DRAFT_OBS: Observable<CaseView> = Observable.of(DRAFT);
 
     let caseResolver: CaseResolver;
@@ -216,13 +220,13 @@ describe('CaseResolver', () => {
       caseResolver
         .resolve(route)
         .then(caseData => {
-          expect(caseData).toBe(DRAFT);
+          expect(caseData).toEqual(DRAFT);
         });
 
       expect(draftService.getDraft).toHaveBeenCalledWith(DRAFT_ID);
       expect(route.paramMap.get).toHaveBeenCalledWith(PARAM_CASE_ID);
       // allows to access private cachedCaseView field
-      expect(caseResolver['cachedCaseView']).toBe(DRAFT);
+      expect(caseResolver['cachedCaseView']).toEqual(DRAFT);
     });
   });
 });