+++ src/shared/components/palette/fixed-radio-list/read-fixed-radio-list-field.component.spec.ts
@@ -31,13 +31,13 @@ describe('ReadFixedRadioListFieldComponent', () => {
   };
 
   describe('Non-persistable readonly fixed radio list field', () => {
-    const CASE_FIELD: CaseField = {
+    const CASE_FIELD: CaseField = <CaseField>({
       id: 'x',
       label: 'X',
       display_context: 'OPTIONAL',
       field_type: FIELD_TYPE,
       value: VALUE
-    };
+    });
     const EMPTY = '';
 
     let fixture: ComponentFixture<ReadFixedRadioListFieldComponent>;
@@ -93,13 +93,13 @@ describe('ReadFixedRadioListFieldComponent', () => {
       FORM_GROUP.addControl(FIELD_ID, control);
       return control;
     };
-    const CASE_FIELD: CaseField = {
+    const CASE_FIELD: CaseField = <CaseField>({
       id: 'x',
       label: 'X',
       display_context: 'OPTIONAL',
       field_type: FIELD_TYPE,
       value: VALUE
-    };
+    });
 
     let fixture: ComponentFixture<ReadFixedRadioListFieldComponent>;
     let component: ReadFixedRadioListFieldComponent;