+++ src/shared/components/palette/phone-uk/read-phone-uk-field.component.spec.ts
@@ -16,13 +16,13 @@ describe('ReadPhoneUKFieldComponent', () => {
   const EMPTY = '';
 
   describe('Non-persistable readonly phone-uk field', () => {
-    const CASE_FIELD: CaseField = {
-      id: FIELD_ID,
-      label: 'X',
-      display_context: 'OPTIONAL',
-      field_type: FIELD_TYPE,
-      value: VALUE
-    };
+  const CASE_FIELD: CaseField = <CaseField>({
+    id: 'x',
+    label: 'X',
+    display_context: 'OPTIONAL',
+    field_type: FIELD_TYPE,
+    value: VALUE
+  });
 
     let fixture: ComponentFixture<ReadPhoneUKFieldComponent>;
     let component: ReadPhoneUKFieldComponent;
@@ -76,13 +76,13 @@ describe('ReadPhoneUKFieldComponent', () => {
       FORM_GROUP.addControl(FIELD_ID, control);
       return control;
     };
-    const CASE_FIELD: CaseField = {
+    const CASE_FIELD: CaseField = <CaseField>({
       id: FIELD_ID,
       label: 'X',
       display_context: 'OPTIONAL',
       field_type: FIELD_TYPE,
       value: VALUE
-    };
+    });
 
     let fixture: ComponentFixture<ReadPhoneUKFieldComponent>;
     let component: ReadPhoneUKFieldComponent;