+++ src/shared/services/index.ts
@@ -16,6 +16,5 @@ export * from './request';
 export * from './search';
 export * from './router';
 export * from './window';
-export * from './workbasket';
 export * from './definitions';
 export * from './search-result/sorting/search-result-view-item-comparator-factory';